ServerShootingStar added
Java console application which is used as a server added. This app sends
an angle which indicates how much the robotic arm should rotate as a
response to the request sent by the C# client app. The resulting angle
is calculated using OpenCV which finds the location of the ball.